Add events for collection updates
Unfortunately, these events say very little and do not contain a
meaningful before/after or diff. There appears to be no way for us to
get this data without reading out the entire join table and the one from
the snapshot, and diffing on that. This is a very expensive operation in
case of a large table.