Use 'item' as link-relation for collection resources
Having an arbitrary link-relation in _embedded is problematic when creating a generic application that consumes a ContentGrid REST API.
A generic application should not have to iterate over all keys in _embedded to find the data in a collection.
There should be the possibility to add additional embedded link-relations later for data related to a collection, without breaking generic applications
that rely on there only being one key in _embedded.

Therefor, we will use the standard 'item' link relation type, which is defined to describe an item in a collection.