Ensure that curies are present on the collection resource
Because the _embedded link-rel itself is no longer a curied link,
spring-data-rest no longer adds the curies link-rel to the top level item;
even though resources inside the embedded collection have link relations
that require curies to be resolved.

Solving this for all collections with EmbeddedWrappers by adding curies
in a ResourceProcessor if hitting the case when spring-data-rest does not add
curies.