Create an implicit relation between `@OneToMany` and a `@Column` on the other side
This is a workaround for the fact that if bi-directional, `@OneToMany` is
always the inverse side of `@ManyToOne`.
However, in contentgrid we only have one-directional relations for now,
with `@OneToMany` the primary side (at least in the model). The other
side can simply not be exposed at all.

The problem with making the `@OneToMany` the inverse side is that
hibernate only considers the primary side when saving changes back to
the database. There currently is no logic to ensure that modifications to
relations are written back to the primary side, so in essence writes
would be discarded.