Add constraints (and validations) to handle required relations before entity delete
The before entity delete is a special case, since the validation is for the ability to cleanly delete the entity.
Those validations are different from the ones used to normally validate the entity (mostly checking that required relations have been severed),
so it needs a separate validation group for that.