Use spring `@ServiceConnection` and testcontainers instead of jdbc-url based testcontainers
It turns out that the jdbc-based testcontainers has a problem where all
connections that use the same jdbc url are shared to the same database
instance. This causes a problem when tests run in parallel, as hibernate
sets up and destroys the schema on every application context start, but
it's all on the same database.

Using the spring `@ServiceConnection` instead ensures that a different
database is started for every time the application context is
(re)created, avoiding any problems with a shared database