Return ResponseEntity directly from `@ControllerAdvice`
Returning a `Problem` instance and letting the normal
`ReturnValueHandler` process the returned value can result in problems
because of content-negotiation.
In that case the problem details are not returned correctly, sometimes receiving a plain
HTTP 406 not acceptable response (when requesting hal-forms only),
or other times the content-type being set to the content-type from the
accept header (in case of `Accept: application/hal+json`).

In both cases, we expect the actual problem details being returned, with
the correct `application/problem+json` content-type.

Returning a `ResponseEntity` directly bypasses the `ReturnValueHandler`
that does content-negotiation, instead writing the response with the
expected data and content-type.