Use new MediaTypeConfigurationCustomizer to customize HalFormsConfiguration
The MediaTypeConfigurationCustomizer is a new way to configure HalFormsConfiguration without creating the whole configuration in one place.
Since we only need to configure hal-forms options on some fields, it is a good candidate to be moved to the MediaTypeConfigurationCustomizer approach.

Also add tests that check that the options are added correctly to relations.