Add test for multiple many-to-many relations to the same entity
There is a bug where multiple many-to-many relations that are referencing the same entity are confused during read operations.
This results in those relations arbitrarily resolving to one of those relations referencing the same entity.

In this test, it could result in the `/orders/xyz/manual-promos` and `/orders/xyz/promos` resolving to either `/promos?orders=xyz` or `/promos?ordersWithManualPromos=xyz`,
depending on which order fields are returned during reflection.