Move logic to create an `OrderSpecifier` to `QuerydslPredicateFactory`
Not all `QuerydslPredicateFactory` implementations support being used
for sorting.

For example, all the `Ordered` predicates do not support being used for
sorting, as it makes little semantic sense to specify a sort on '<property> is
less than' as opposed to a sort on '<property>' itself.

Similarly, the `Text.StartsWith` predicates also don't make sense to be
used for sorting.