Only handle jackson exceptions when wrapped inside HttpMessageNotReadableException
Handling Jackson exceptions generically means that they can also be triggered by other parts of the request-response cycle,
including during response serialization.
Sending a 400 error for things that are not related to actual input message parsing is bad form
and makes it more difficult to diagnose the actual issue (compared to having a HTTP 500 happen for an unhandled exception)

Due to the setup for `@ExceptionHandler` methods from Spring, all cases for an exception need to be handled in one method.
There is no way to automatically dispatch different methods based on the nested cause of an exception.