Always have `total_items_estimate` present
An exact count is also valid for an "estimate"; this way applications
that don't care about the exact count and always want to work off a kind
of estimate don't have to switch between the `total_items_exact` and
`total_items_estimate` fields.

Additionally, make `total_items_estimate` a number instead of a string.
Application code can have valid reasons to treat the estimate as a
number, for example formatting the number in the native language of the
user.
Making it a string just makes that more difficult for the developer.
Any person interpreting a clearly labelled estimate as an exact number
will get what they deserve when that assumption breaks down.