Configure test application with short datasource timeout
On test completion, the database is already shutdown before the hibernate shutdown runs.
The result is that hibernate waits until timeout (default 30s) for a connection to run
cleanup commands (table drops). Setting the timer to a shorter value ensures that our
tests complete much faster, instead of the process hanging for 30 seconds before terminating
succesfully