Add request URI as context to cursor encode/decode
The UriComponents for the request URI (without page/pageSize parameters)
are passed as context to encode and decode cursors.

This can be used to add a checksum into the cursor to ensure that
cursors are not inadvertently re-used with different parameters.
A cursor is only valid for a particular set of request parameters,
otherwise the results of using a cursor are undefined.