Ensure that HalConfiguration MediaTypeConfigurationCustomizers are always run
spring-data-rest is not aware of the existence of MediaTypeConfigurationCustomizer, so it does not apply the configuration when pulling the HalConfiguration bean directly.
This leads to the configuration settings in HalConfiguration inconsistently being applied, depending on if a setting is used in spring-hateoas or spring-data-rest.

Always applying all customizers directly in a BeanPostProcessor ensures that the HalConfiguration bean itself has all the customizations directly baked-in,
and does not require all consumers to be aware of MediaTypeConfigurationCustomizer