Move contentgrid automations link relations to their own URI namespace
Create a separate namespace for automations, so the `cg:*` link
relations are not being reused in a different context at all.

Also move automations configuration to the module itself, and have the
autoconfiguration only perform imports.
This allows a cleaner extension and the possibility to test automations
separately from the rest.