Fix event publication test for spring boot 3.4
It is not supported to *set* the id of an entity before saving.
This results in an optimistic locking failure.

Because hibernate assumes that an object with an ID was already loaded
from a database, it throws `StaleObjectStateException: Row was updated or deleted by another transaction (or unsaved-value mapping was incorrect)`