Prevent clashes between m2m relations with the same name
When multiple many-to-many relations with the same name (from different
source entities) target the same entity; the GET relation endpoint did
not redirect with the correct collection filter.

It should be considered that multiple `@ManyToMany(mappedBy="...")`
annotations with the same `mappedBy` value can legally be present when
they refer to different entities on the other side.